update all circle charts to use new d3 circle chart and legends
