+++ .env.dev
@@ -1,3 +1,3 @@
-GATSBY_HASURA_URI="https://hasura-sandbox.app.cloud.gov/v1/graphql"
+GATSBY_HASURA_URI="https://hasura-dev.app.cloud.gov/v1/graphql"
 GTM_ID="GTM-NCRF98R"
 isShutdown=false
