+++ src/components/data-viz/svg/Label.js
@@ -4,8 +4,10 @@ export const Label = ({ data, key, fill, fontSize, isClickable }) => {
   console.log('Label data: ', data)
   const styles = {
     default: {
-      cursor: 'pointer',
       fontSize: fontSize
+    },
+    clickable: {
+      cursor: 'pointer',
     }
   }
   return (
@@ -13,7 +15,7 @@ export const Label = ({ data, key, fill, fontSize, isClickable }) => {
       key={`text__${ key }`}
       data={data}
       fill={fill}
-      style={isClickable ? styles.default : {}}
+      style={isClickable ? { ...styles.default, ...styles.clickable } : { ...styles.default }}
     >
       {data}
     </text>
