+++ src/components/sections/ExploreData/DetailCards.js
@@ -145,7 +145,7 @@ const useStyles = makeStyles(theme => ({
       maxWidth: '100%',
     },
     '& .cardContent__Revenue': {
-      gridTemplateRows: '185px 660px 660px',
+      gridTemplateRows: '185px 720px 600px',
     },
     '& .cardContent__Disbursements': {
       gridTemplateRows: '185px 715px 660px',
