+++ src/pages/explore/index.mdx
@@ -213,7 +213,7 @@ tags:
                             <DetailCards>     
                               <ProductionDetailTrends />
                               <ProductionLandCategory title="Production over time"  />        
-                              <ProductionTopLocations title="Top locations" vertical maxLegendWidth='800px' />   
+                              <ProductionTopLocations title="Top locations" />   
                               <ProductionCountyMap />
                             </DetailCards>
                           </Grid>
